bf16 + fp32 master weights, memmapped data, train.py
MasterWeightOptimizer keeps bf16 model params synced to fp32 masters
stepped by Adam; the precision test shows plain-bf16 Adam stalling
behind while bf16+masters tracks fp32 to a ~0.006 final-loss gap, and
verifies params stay bitwise equal to master.to(bf16). Pipeline p2p
buffers are now dtype-aware.

TokenShard memmaps a uint16 shard (prepare_openwebtext.py writes one);
batches draw at offsets seeded by (seed, step, dp_rank) so a replica's
ranks agree on the batch with zero communication while DP replicas
diverge. train.py composes every axis behind flags; 8-way
dp2 x pp2 x tp2 interleaved+SP+recompute+bf16 drops loss 5.57 -> 3.51
on a real byte-level shard.

+++ dst/model.py
@@ -274,7 +274,10 @@ class GPT(nn.Module):
 
         loss = None
         if targets is not None:
-            loss = F.cross_entropy(logits.reshape(-1, logits.shape[-1]), targets.reshape(-1))
+            # Loss in fp32 regardless of model dtype (no-op when fp32).
+            loss = F.cross_entropy(
+                logits.reshape(-1, logits.shape[-1]).float(), targets.reshape(-1)
+            )
         return logits, loss
 
 
@@ -362,7 +365,9 @@ class GPTStage(nn.Module):
         logits = self.lm_head(x)[..., : self.config.vocab_size]
         loss = None
         if targets is not None:
-            loss = F.cross_entropy(logits.reshape(-1, logits.shape[-1]), targets.reshape(-1))
+            loss = F.cross_entropy(
+                logits.reshape(-1, logits.shape[-1]).float(), targets.reshape(-1)
+            )
         return logits, loss
